Move orchestration functions out of state/index.js to callers
selectNode/deselectNode → navigation.js
probeNode → probe-exec.js
ejectIce/disableIce/rebootIce → ice.js
readNode/lootNode/reconfigureNode/rebootNode/completeReboot → node-orchestration.js (new)

state/index.js now contains only: initState, getState, mutate, getVersion,
ALERT_ORDER, revealNeighbors, accessNeighbors, endRun, buyExploit,
isIceVisible, and serialization. No more mixed mutation+event functions.
