+++ js/navigation.js
@@ -6,7 +6,9 @@
 // (exploit-exec, probe-exec, ice) subscribe to that event to perform their
 // own cleanup — navigation.js does not need to know about them.
 
-import { selectNode, deselectNode, getState } from "./state.js";
+import { getState } from "./state.js";
+import { setSelectedNode } from "./state/game.js";
+import { setNodeVisible } from "./state/node.js";
 import { emitEvent, E } from "./events.js";
 
 /**
@@ -15,8 +17,29 @@ import { emitEvent, E } from "./events.js";
  * @param {string} nodeId
  */
 export function navigateTo(nodeId) {
-  const isNewNode = getState().selectedNodeId !== nodeId;
-  selectNode(nodeId);
+  const s = getState();
+  const isNewNode = s.selectedNodeId !== nodeId;
+
+  const node = s.nodes[nodeId];
+  if (node?.rebooting) {
+    emitEvent(E.LOG_ENTRY, { text: `${node.label}: node is rebooting.`, type: "error" });
+    return;
+  }
+  setSelectedNode(nodeId);
+
+  // Traversal: selecting a revealed ("???") node adjacent to any accessible node makes it
+  // accessible. This is how the player explores deeper into the network.
+  if (node && node.visibility === "revealed") {
+    const hasAccessibleNeighbor = (s.adjacency[nodeId] || []).some(
+      (nid) => s.nodes[nid]?.visibility === "accessible"
+    );
+    if (hasAccessibleNeighbor) {
+      setNodeVisible(nodeId, "accessible");
+      emitEvent(E.NODE_REVEALED, { nodeId, label: node.label });
+      emitEvent(E.LOG_ENTRY, { text: `[NODE] ${node.label}: signal traced. Node accessible.`, type: "info" });
+    }
+  }
+
   if (isNewNode) emitEvent(E.PLAYER_NAVIGATED, { nodeId });
 }
 
@@ -24,6 +47,6 @@ export function navigateTo(nodeId) {
  * Deselect the current node — clears the selection and notifies subscribers.
  */
 export function navigateAway() {
-  deselectNode();
+  setSelectedNode(null);
   emitEvent(E.PLAYER_NAVIGATED, { nodeId: null });
 }
